backend/requirements.txt
Django==4.2.4
djangorestframework==3.14.0
django-cors-headers==4.2.0
drf-spectacular==0.26.4